+++ mitra-models/src/profiles/types.rs
@@ -141,7 +141,7 @@ pub enum IdentityProofType {
     FepC390JcsBlake2Ed25519Proof, // MitraJcsEd25519Signature2022
     FepC390JcsEip191Proof, // MitraJcsEip191Signature2022
     FepC390LegacyJcsEddsaProof, // jcs-eddsa-2022
-    FepC390JcsEddsaProof, // eddsa-jcs-2022
+    FepC390EddsaJcsNoCiProof, // was used for incorrect eddsa-jcs-2022 proofs
 }
 
 impl From<&IdentityProofType> for i16 {
@@ -152,7 +152,7 @@ impl From<&IdentityProofType> for i16 {
             IdentityProofType::FepC390JcsBlake2Ed25519Proof => 3,
             IdentityProofType::FepC390JcsEip191Proof => 4,
             IdentityProofType::FepC390LegacyJcsEddsaProof => 5,
-            IdentityProofType::FepC390JcsEddsaProof => 6,
+            IdentityProofType::FepC390EddsaJcsNoCiProof => 6,
         }
     }
 }
@@ -167,7 +167,7 @@ impl TryFrom<i16> for IdentityProofType {
             3 => Self::FepC390JcsBlake2Ed25519Proof,
             4 => Self::FepC390JcsEip191Proof,
             5 => Self::FepC390LegacyJcsEddsaProof,
-            6 => Self::FepC390JcsEddsaProof,
+            6 => Self::FepC390EddsaJcsNoCiProof,
             _ => return Err(DatabaseTypeError),
         };
         Ok(proof_type)