+++ src/activitypub/actors/attachments.rs
@@ -185,7 +185,6 @@ pub fn parse_identity_proof_fep_c390(
             ).map_err(|_| ValidationError("invalid identity proof"))?;
             IdentityProofType::FepC390JcsEip191Proof
         },
-        // TODO: ProofType::EddsaJcsSignature
         ProofType::JcsEddsaSignature => {
             let did_key = signer.as_did_key()
                 .ok_or(ValidationError("unexpected DID type"))?;
@@ -199,7 +198,11 @@ pub fn parse_identity_proof_fep_c390(
                 &signature_data.proof_config,
                 &signature_data.signature,
             ).map_err(|_| ValidationError("invalid identity proof"))?;
-            IdentityProofType::FepC390JcsEddsaProof
+            IdentityProofType::FepC390LegacyJcsEddsaProof
+        },
+        ProofType::EddsaJcsSignature => {
+            // eddsa-jcs-2022 identity proofs are temporarily rejected
+            return Err(ValidationError("eddsa-jcs-2022 cryptosuite is not supported"));
         },
         _ => return Err(ValidationError("unsupported signature type")),
     };